[StyleCop] Fix all the warnings on ApplicationInsights.Tests
