+++ tests/Microsoft.Bot.Builder.ApplicationInsights.Tests/BotTelemetryClientTests.cs
@@ -1,15 +1,15 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.ApplicationInsights;
-using Microsoft.ApplicationInsights.DataContracts;
-using System.Collections.Generic;
 using System;
-using Microsoft.ApplicationInsights.Extensibility;
-using Moq;
+using System.Collections.Generic;
+using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Channel;
+using Microsoft.ApplicationInsights.DataContracts;
+using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.Bot.Builder.ApplicationInsights;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Moq;
 
 namespace Microsoft.Bot.Builder.Integration.ApplicationInsights.Tests
 {
@@ -40,11 +40,12 @@ namespace Microsoft.Bot.Builder.Integration.ApplicationInsights.Tests
 
                 var botTelemetryClient = new BotTelemetryClient(telemetryClient);
             }
+
             [TestMethod]
             public void OverrideTest()
             {
                 var telemetryClient = new TelemetryClient();
-                var myTelemetryClient = new MyBotTelemetryClient(telemetryClient);
+                var botTelemetryClient = new MyBotTelemetryClient(telemetryClient);
             }
         }
 
@@ -68,15 +69,21 @@ namespace Microsoft.Bot.Builder.Integration.ApplicationInsights.Tests
             [TestMethod]
             public void TrackAvailabilityTest()
             {
-                _botTelemetryClient.TrackAvailability("test", DateTimeOffset.Now, new TimeSpan(1000), "run location", true,
-                    "message", new Dictionary<string, string>() { { "hello", "value" } }, new Dictionary<string, double>() { { "metric", 0.6 } });
+                _botTelemetryClient.TrackAvailability(
+                    "test",
+                    DateTimeOffset.Now,
+                    new TimeSpan(1000),
+                    "run location",
+                    true,
+                    "message",
+                    new Dictionary<string, string>() { { "hello", "value" } },
+                    new Dictionary<string, double>() { { "metric", 0.6 } });
 
                 _mockTelemetryChannel.Verify(tc => tc.Send(It.Is<AvailabilityTelemetry>(t => t.Name == "test")));
                 _mockTelemetryChannel.Verify(tc => tc.Send(It.Is<AvailabilityTelemetry>(t => t.Message == "message")));
                 _mockTelemetryChannel.Verify(tc => tc.Send(It.Is<AvailabilityTelemetry>(t => t.Properties["hello"] == "value")));
                 _mockTelemetryChannel.Verify(tc => tc.Send(It.Is<AvailabilityTelemetry>(t => t.Metrics["metric"] == 0.6)));
             }
-
 
             [TestMethod]
             public void TrackEventTest()
@@ -121,44 +128,6 @@ namespace Microsoft.Bot.Builder.Integration.ApplicationInsights.Tests
                 _mockTelemetryChannel.Verify(tc => tc.Send(It.Is<TraceTelemetry>(t => t.SeverityLevel == SeverityLevel.Critical)));
                 _mockTelemetryChannel.Verify(tc => tc.Send(It.Is<TraceTelemetry>(t => t.Properties["foo"] == "bar")));
             }
-        }
-    }
-    public class MyBotTelemetryClient : BotTelemetryClient
-    {
-        public MyBotTelemetryClient(TelemetryClient telemetryClient)
-            : base(telemetryClient)
-        {
-            
-        }
-
-        public override void TrackDependency(string dependencyTypeName, string target, string dependencyName, string data, DateTimeOffset startTime, TimeSpan duration, string resultCode, bool success)
-        {
-            base.TrackDependency(dependencyName, target, dependencyName, data, startTime, duration, resultCode, success);
-        }
-
-        public override void TrackAvailability(string name, DateTimeOffset timeStamp, TimeSpan duration, string runLocation, bool success, string message = null, IDictionary<string, string> properties = null, IDictionary<string, double> metrics = null)
-        {
-            base.TrackAvailability(name, timeStamp, duration, runLocation, success, message, properties, metrics);
-        }
-
-        public override void TrackEvent(string eventName, IDictionary<string, string> properties = null, IDictionary<string, double> metrics = null)
-        {
-            base.TrackEvent(eventName, properties, metrics);
-        }
-
-        public override void TrackException(Exception exception, IDictionary<string, string> properties = null, IDictionary<string, double> metrics = null)
-        {
-            base.TrackException(exception, properties, metrics);
-        }
-
-        public override void TrackTrace(string message, Severity severityLevel, IDictionary<string, string> properties)
-        {
-            base.TrackTrace(message, severityLevel, properties);
-        }
-
-        public override void Flush()
-        {
-            base.Flush();
         }
     }
 }
